move insights-dashboard page to campaigns
