+++ apps/frontend/app/(website)/campaigns/[campaignId]/view/[viewId]/components/CardTile.tsx
