+++ apps/frontend/app/(website)/campaigns/[campaignId]/view/[viewId]/components/DynamicLineChart.tsx
