+++ apps/frontend/app/(website)/campaigns/[campaignId]/view/[viewId]/mocks/card-revenue.json
