+++ apps/frontend/app/(website)/campaigns/[campaignId]/view/[viewId]/mocks/chart-netlify.json
