+++ apps/frontend/app/(website)/campaigns/[campaignId]/view/[viewId]/mocks/table-heroes.json
