+++ apps/frontend/app/(website)/campaigns/[campaignId]/view/[viewId]/types.ts
