+++ apps/frontend/app/(website)/campaigns/[campaignId]/view/[viewId]/widgets/HeroesTable.tsx
@@ -1,8 +1,8 @@
-import { TableTile } from "@/app/(website)/insights-dashboard/components/TableTile";
+import { TableTile } from "@/app/(website)/campaigns/[campaignId]/view/[viewId]/components/TableTile";
 import type {
   MigrationHeroData,
   TransformerType,
-} from "@/app/(website)/insights-dashboard/types";
+} from "@/app/(website)/campaigns/[campaignId]/view/[viewId]/types";
 import { Clock, Code, Plus, Trash } from "@phosphor-icons/react";
 import { User } from "../components/User";
