+++ apps/frontend/app/(website)/campaigns/layout.tsx
@@ -13,9 +13,9 @@ const Layout = ({ children }: { children: React.ReactNode }) => {
   return (
     <Theme>
       <QueryClientProvider client={client}>
-        <div className="h-[calc(100vh-76px)] mt-[76px] flex">
+        <div className="h-[100vh] pt-[calc(var(--header-height))] flex">
           {isSidebarActive && <Sidebar />}
-          <div className="flex-grow">{children}</div>
+          <div className="flex-grow h-full overflow-y-auto">{children}</div>
         </div>
       </QueryClientProvider>
     </Theme>
